Moves methods to clean t_commands into a new file
